Add a `stage locate-bun-depl-pkg` command
