+++ cmd/forklift/stage/store.go
@@ -1,8 +1,6 @@
 package stage
 
 import (
-	"fmt"
-
 	"github.com/pkg/errors"
 
 	"github.com/PlanktoScope/forklift/internal/app/forklift"
@@ -12,15 +10,7 @@ var errMissingStore = errors.Errorf(
 	"you first need to stage a pallet, e.g. with `forklift plt stage`",
 )
 
-func getStageStore(wpath string, ensureWorkspace bool) (*forklift.FSStageStore, error) {
-	if ensureWorkspace {
-		if !forklift.Exists(wpath) {
-			fmt.Printf("Making a new workspace at %s...", wpath)
-		}
-		if err := forklift.EnsureExists(wpath); err != nil {
-			return nil, errors.Wrapf(err, "couldn't make new workspace at %s", wpath)
-		}
-	}
+func getStageStore(wpath string) (*forklift.FSStageStore, error) {
 	workspace, err := forklift.LoadWorkspace(wpath)
 	if err != nil {
 		return nil, err
